correction

+++ DL/LSTM_model_v2.py
@@ -1622,8 +1622,8 @@ class LSTM_model(DL):
         if plotting==True:
             import matplotlib.dates as md
             a = np.round(cv, 2)
-            up =int(np.max(y_realF)) + int(np.max(y_realF)/4)
-            low = int(np.min(y_realF)) - abs(int(np.min(y_realF)/4))
+            up =np.max(y_realF) + np.max(y_realF)/4
+            low = np.min(y_realF) - abs(np.min(y_realF)/4)
             fig,ax = plt.subplots(1,1)
             plt.ylim(low, up)
             plt.plot(y_realF, color='black', label='Real')
